Adding an implementation of a plugin system.
The loader.c file contains the core of our plugin machinery. It uses
`dlopen()' to find and open a dynamic library (.so file) and using
slc_plugin_ctx, it is possible to store the loaded plugin into a
hashtable and search for it when a message arrives from the user via
xmpp.

An example of how to use it was added too. It is in the test-plugin.c
file.

A ready to use hashtable implementation was used here. Thank you Petri
Lehtinen <petri@digip.org> for this =D