Adding a method to load plugins from a file.
Now it is not necessary to load plugins by hardcoding a call
to `slc_plugin_ctx_load()` in the main.c file. Now you can just
write a config file with a dynamic library per line and pass
it to the --plugins-file argument.

My real plan is to load it through an interactive shell, but
I am still working on this and making tests with hardcoded
stuff is not good at all.