Adding a client/server based shell.
This shell connects to the server running *inside* bitu
through an Unix socket. It has its internal commands but
the main goal is to manipulate data in the running bitu
instance.

It is in my mind to write a little language to manage
plugins and xmpp interaction in general. Maybe, I
should just use guile or some other scheme
implementation.