Collecting parameters from the message sent by the user.
Now, the message is broken in tokens (separed by spaces) and
the first one is considered the command and the rest is stored
in a `char **` variable and are considered parameters.

Nothing is being done with these parameters in this commit. The
next problem to be solved is how to pass this info to the
plugin callback