Creating a struct to hold main app stuff.
The `bitu_app_t` struct was created to hold all important data
of the application. A single instance of this type is created
for the whole application but nobody will die if someone try to
use it in another place. No constructor or destructor was
writen for this stuff.

The next step is to pass this instance to the `bitu_server_run()`
function. To allow it to use things like the logger and in a
not so far future, the config system.