Heavy improvements in the internal server and the shell.
Now, the `bitu_server_run()` function receives the main app context with
log, xmpp and plugin context structs. All printfs were replaced by the
`ta_log_*` calls.

But the main change is the hashtable that holds server commands. Now,
a command must be registered in this hashtable and when the server
receives a command from the client, a lookup is performed in this hash.
If found, it is called with the main app context as extra param. The
command params are extracted using `bitu_util_extract_params()` and
then passed to the callback too.

Two commands were implemented in this commit. The `load` and `unload`
ones.