Adding a new struct to abstract the server stuff.
The socket connection code was moved to the `bitu_server_connect()`
method and there is more error handling than before. The method
`bitu_server_run()` contains only the accept and send/recv stuff in
a loop.

The next steps are the use of select() and moving accept and send/recv
to a _recv, _send method.