Adding a config file reader and a sample config file.
Now, connection info, plugins to be loaded at the start of the program
and some other stuff can be stored in a config file. At this time,
only read operations are performed in this file. Soon we'll write the
module list loaded from the shell to this config file.

ps.: Some forgotten headers were added to include/bitu/Makefile.am in
this commit too.