Heavy changes in the implementation of local server and client.
First of all, the interactive shell is being broken in this
commit. The first problem that causes this issue is the select()
timeout. In previous versions I was not using select and sometimes
recv was blocking (when it shouldn't =P). The main issue was almost
fixed but now, the interactive shell doesn't work because the timeout
added to select.

Another thing I did is to catch the SIGPIPE signal. It was breaking
the server when the client goes down.

The `bitu_server_exec_cmd()' has its prototype changed too. Now it
returns an answer of the executed command (if it retorns something, of
course). And the `bitu_server_exec_cmd_line()' method was added.

Now, the server sends mesages using `bitu_server_send()' method
instead of raw `send()'. It tests how much data was sent.

In the client side, the command struct was changed too. Now it has the
number of parameters, the help command was purged. It should become
from the server that really knows all other commands, the `recv()'
call was encapsulated in an `almost fault tolerant' and `select()
enabled' function: `bitu_shell_recv()'