Implementing the `set-log-file' command in the server.
Now it is possible to save all log messages to a file. If you change
this stuff while server is running you will not need to restart the
bot. It will only stop to write in the previous file and append stuff
to the new one.

If there is any problem when opening the new log file, the operation
will not be completed and the old log file will continue to be used.