Adding the base of the new connection manager
This connection manager is powered by a producer/consumer style queue,
fed by the connected clients and consumed in a new thread started in the
main.c file.

In this change, I also made the `bitu_transport_t` type an opaque struct
and provided all the methods needed to access things that needs to be
accessed.