Using our local server as any other regular transport
This commit makes the command/plugin evaluation stop working properly cause
the server is not responsible for executing them anymore. This job will be
transfered to another component.

Now our server is just another regular transport and is spinned up when the
user sets the transport to "file:///path/to/socket" in the config file or in
the command line arguments.

Every time a new client connects, the hashtable server->clients receives
an instance of a `bitu_client_t` object and each client has an id, built
using the libuuid. This way, we can send messages back to the specific
user that asked for a command/plugin.