Adding the "app" component, to organize all the other components
Now, the configuration read function does not create the custom conf_entry_t
instance anymore. It creates a bitu_command_instance, with the transport and
"from" variables set to NULL.

Also, the "transport" command syntax changed to "transport add" and its not
a special case anymore in the command exec loop.

Currently, we are not aware of plugins yet, we are just executting commands,
but probably the next patch will make it work again