Adding commands to add/list/rm/connect/disconnect transports
To make it possible, two more functions were added to the transport API,
the disconnect() and is_running() ones. It means that now the transport
interface requires the implementer of a new one to set callbacks for
these both things.

In the app side, the sub commands for the "transport" command were added
with a nice error check and reasonable error messages.

Both local and xmpp transports were updated to reflect the above
changes. The irc one which is not in the same state of quality is left
without any improvements. So it is definitely broken and cant be used
right now. :(

Closes #6