Allowing the plugin to optionally handle the entire command line
Instead of just receiving the plugin name executing it with poorly
parsed arguments, this commit allows a plugin to say if it's able to
handle a command line or not with the new "plugin->match" callback.

The "plugin->num_params" callback was removed and now the
"plugin->execute" callback takes the a bitu_command_t instance as the
only argument and the plugin writer can access all the data using the
command API.

The method "bitu_app_exec_command()" prioritizes the internal commands
when receiving a new command to parse. If no internal command can be
found, we try to find a plugin that matches the command received.

Closes #4