Making the bitu_transport_run() not start a new thread
To make it easier to debug transports, I haveve added a new program that
just creates a new transport, connect and run it. This way, we do not
actually start any thread and if the transport fails, I will have my
precious backtrace.

Another change is that now all the transports must test if it was
possible to queue a command or not. If it is not, it just returns
something to say that the command has failed.